+++ ecommerce-product-service-api/src/main/java/com/ecommerce/product/about/AboutController.java
@@ -1,6 +1,6 @@
 package com.ecommerce.product.about;
 
-import com.ecommerce.product.representation.about.AboutRepresentation;
+import com.ecommerce.product.sdk.representation.about.AboutRepresentation;
 import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
