+++ ecommerce-product-service-api/src/main/java/com/ecommerce/product/product/ProductApplicationService.java
@@ -1,7 +1,7 @@
 package com.ecommerce.product.product;
 
-import com.ecommerce.product.command.product.CreateProductCommand;
-import com.ecommerce.product.command.product.UpdateProductNameCommand;
+import com.ecommerce.product.sdk.command.product.CreateProductCommand;
+import com.ecommerce.product.sdk.command.product.UpdateProductNameCommand;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
