+++ ecommerce-product-service-api/src/main/java/com/ecommerce/product/product/ProductRepresentationService.java
@@ -1,6 +1,6 @@
 package com.ecommerce.product.product;
 
-import com.ecommerce.product.representation.product.ProductSummaryRepresentation;
+import com.ecommerce.product.sdk.representation.product.ProductSummaryRepresentation;
 import com.ecommerce.shared.utils.PagedResource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
