+++ ecommerce-product-service-sdk/src/main/java/com/ecommerce/product/sdk/event/category/CategoryCreatedEvent.java
@@ -1,4 +1,4 @@
-package com.ecommerce.product.event.category;
+package com.ecommerce.product.sdk.event.category;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
